Changes in/Update Migrations/Schema 3. Improve Compatibility.

+++ db/migrate/20180317204452_create_cust_g_funcs.rb
@@ -1,14 +1,13 @@
 class CreateCustGFuncs < ActiveRecord::Migration[5.0]
   def change
     create_table :cust_g_funcs do |t|
-
       t.belongs_to :cust_func, foreign_key: true, null: false
       t.belongs_to :game, foreign_key: true, null: false, index: true
       
-      t.decimal :parA
-      t.decimal :parB
-      t.decimal :parC
-      t.decimal :parD
+      t.float :parA
+      t.float :parB
+      t.float :parC
+      t.float :parD
       
       t.timestamps
     end
